requirements.txt
astroid==2.2.5
awscli==1.16.144
botocore==1.12.134
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
Django==2.2.3
django-http-method==1.0.2
docker==4.2.0
docutils==0.14
get==2019.4.13
gitdb==4.0.4
GitPython==3.1.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
pipenv==2018.11.26
post==2019.4.13
public==2019.4.13
pyasn1==0.4.5
pylint==2.3.1
python-dateutil==2.8.0
python-twitch==1.3.0
python-twitch-client==0.6.0
pytz==2019.1
PyYAML==3.13
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
smmap==3.0.2
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.2
virtualenv==16.4.3
virtualenv-clone==0.5.3
websocket-client==0.57.0
wrapt==1.11.2
